+++ rocprim/include/rocprim/type_traits_functions.hpp
@@ -1,4 +1,4 @@
-// Copyright (c) 2017-2025 Advanced Micro Devices, Inc. All rights reserved.
+// Copyright (c) 2017-2026 Advanced Micro Devices, Inc. All rights reserved.
 //
 // Permission is hereby granted, free of charge, to any person obtaining a copy
 // of this software and associated documentation files (the "Software"), to deal
@@ -273,8 +273,7 @@ struct is_tuple_of_references<::rocprim::tuple<Args...>>
 private:
     template<size_t Index>
     ROCPRIM_HOST_DEVICE
-    static constexpr
-        typename std::enable_if<(Index < sizeof...(Args)), bool>::type is_tuple_of_references_impl()
+    static constexpr std::enable_if_t<(Index < sizeof...(Args)), bool> is_tuple_of_references_impl()
     {
         using tuple_t   = ::rocprim::tuple<Args...>;
         using element_t = ::rocprim::tuple_element_t<Index, tuple_t>;
@@ -283,7 +282,7 @@ private:
 
     template<size_t Index>
     ROCPRIM_HOST_DEVICE
-    static constexpr typename std::enable_if<(Index == sizeof...(Args)), bool>::type
+    static constexpr std::enable_if_t<(Index == sizeof...(Args)), bool>
         is_tuple_of_references_impl()
     {
         return true;
@@ -543,14 +542,33 @@ template<typename T>
 struct word_type<const volatile T> : word_type<T>
 {};
 
-} // namespace detail
-
-namespace detail
-{
 template<typename Destination, typename Source>
 constexpr bool is_valid_bit_cast
     = sizeof(Destination) == sizeof(Source) && std::is_trivially_copyable<Destination>::value
       && std::is_trivially_copyable<Source>::value;
+
+/// \brief Utility to apply functions over multiple types.
+template<class... Ts>
+struct variadic_list
+{
+    /// \brief Applies predicate `f` over `Ts` and returns if any is true.
+    template<class F>
+    static constexpr bool any(F f)
+    {
+        return (f(Ts{}) || ...);
+    }
+
+    /// \brief Applies predicate `f` over `Ts` type and returns if all is true.
+    template<class F>
+    static constexpr bool all(F f)
+    {
+        return (f(Ts{}) && ...);
+    }
+};
+static_assert(variadic_list<short, int>::all([](auto t)
+                                             { return std::is_integral_v<decltype(t)>; }));
+static_assert(variadic_list<float, int>::any([](auto t)
+                                             { return std::is_integral_v<decltype(t)>; }));
 } // namespace detail
 
 END_ROCPRIM_NAMESPACE