+++ test/rocprim/test_utils_data_generation.hpp
@@ -1,4 +1,4 @@
-// Copyright (c) 2021-2025 Advanced Micro Devices, Inc. All rights reserved.
+// Copyright (c) 2021-2026 Advanced Micro Devices, Inc. All rights reserved.
 //
 // Permission is hereby granted, free of charge, to any person obtaining a copy
 // of this software and associated documentation files (the "Software"), to deal
@@ -147,6 +147,37 @@ public:
         return bfloat16(std::numeric_limits<float>::signaling_NaN());
     };
 };
+
+template<>
+struct numeric_limits<__hip_bfloat16> : std::numeric_limits<__hip_bfloat16>
+{
+public:
+    static inline __hip_bfloat16 max()
+    {
+        return __hip_bfloat16(std::numeric_limits<float>::max() * 0.998);
+    };
+    static inline __hip_bfloat16 min()
+    {
+        return __hip_bfloat16(std::numeric_limits<float>::min());
+    };
+    static inline __hip_bfloat16 lowest()
+    {
+        return __hip_bfloat16(std::numeric_limits<float>::lowest() * 0.998);
+    };
+    static inline __hip_bfloat16 infinity()
+    {
+        return __hip_bfloat16(std::numeric_limits<float>::infinity());
+    };
+    static inline __hip_bfloat16 quiet_NaN()
+    {
+        return __hip_bfloat16(std::numeric_limits<float>::quiet_NaN());
+    };
+    static inline __hip_bfloat16 signaling_NaN()
+    {
+        return __hip_bfloat16(std::numeric_limits<float>::signaling_NaN());
+    };
+};
+
 // End of extended numeric_limits
 
 END_ROCPRIM_NAMESPACE
@@ -334,8 +365,9 @@ auto make_distribution(U min, V max)
 {
     if constexpr(rocprim::is_floating_point<T>::value)
     {
-        using dis_type = typename std::conditional<std::is_same<rocprim::half, T>::value
-                                                       || std::is_same<rocprim::bfloat16, T>::value,
+        using dis_type = typename std::conditional<std::is_same_v<rocprim::half, T>
+                                                       || std::is_same_v<rocprim::bfloat16, T>
+                                                       || std::is_same_v<__hip_bfloat16, T>,
                                                    float,
                                                    T>::type;
         return std::uniform_real_distribution<dis_type>(static_cast<dis_type>(min),